requirements.txt
Django
djangorestframework
psycopg2-binary
Pillow
djangorestframework
django-decouple
django-filter
drf_yasg
